extract to HeaderDesktop component
